requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
svgwrite>=1.4.0
Flask>=2.0.0
Werkzeug>=2.0.0
potrace>=0.2.0
